Temporary hack.
For org to find out what the tags of a heading are, the heading must be
visible. I don't know any way to make the heading visible and then restore the
previous visibility of the buffer when my function ends. For the time being, the
function that finds out whether the heading is errand then leaves the buffer
visibility in a temporary state. The function that exports the errand task now
works.