Addition of doc-mode-split-keywords that splits the documentation using the explicit keywords.
The previous version, splitting using only \ and @, was problematic when
inserting \em or other formatting elements not meants to be put at the beginning
of a block. Any use of keyword not explicitely indicated in
doc-mode-template-keywords will not be handled.

Also add doc-mode-keywords-regexp that returns a regexp matching a keyword.