Get rid of the konix/org-agenda-tag-filter-preset option.
I often want to filter agenda's entries with a or pattern, showing
for instances tasks that may be done at home (@home) or at a
computer.  To do so, the auto-exclude function seems to be the way
to go.

The feature of konix/org-agenda-tag-filter-preset is then replaced
by a more general konix/org-agenda-tag-filter-contexts feature
that remove any context (tag beginning with @) not in
konix/org-agenda-tag-filter-contexts.

The feature behaves like konix/org-agenda-tag-filter-preset in the sense that
the value of the feature is taken into account when any agenda is created.

To do so, org-agenda is adviced to perform the filter automatically like if /
Ret was pressed at the beginning of each agenda.

The hotkeys are updated to follow the change.