Try to launch the per_site_user_agent.sh synchronously.
This is done to avoid the request to be sent to the server before
the user agent has been changed.

Actually, the per_site_user_agent.sh changes the user agent in a
asynchronous way. Even if it is launched at the earliest time, the
request with uzbl user agent is made before it has been changed by
the script.

I am not sure whether doing a sync_spawn will insure the script
will be terminated before the page is requested and then the user
agent will be correctly set, but it is worth trying.