Simplify a lot the life status of an entry
Now, there are only TODO -> NEXT -> DONE. The waiting status is
put into a tag. This reduces the number of reports and allows more
flexibility in the description of the status of a task.

- A general TODO entry is marked TODO
- The next thing to do is marked NEXT
- A waiting entry that blocks the projects is a NEXT :WAIT:
- A delegated entry that blocks the projects is a NEXT :DELEGATED:
- A waiting entry that does not block the projects is a TODO :WAIT:
- A delegated entry that does not block the projects is a TODO :DELEGATED:

A report has been added to be sure a blocking waiting entry is not
making a project stuck without me noticing.