Attempt to allow a project not to be partially stuck
If project A contains project B and project B has task C as next
action, may be project A may have another task than C as next
action. The fact that it waits for the task C makes it partially stuck.

Now, a project is considered stuck if it has no next task in its
direct children.

This makes the workflow of B and A more independant and yet allows
to links them via the next todo state.

In the case of the A, B, C example, A would be considered
stuck. To "unstuck" it, one would have to put the next state to B
or another task D considered relevant.